+++ tests/integration/examples/test_examples.py
@@ -6,14 +6,10 @@ Test examples
 
 
 def test_register_asset():
-    from examples.register_asset import main
-    main()
-
-def test_register_memory_asset():
-    from examples.register_memory_asset import main
+    from examples.register_asset_squid import main
     main()
 
 
-def test_surfer_memory_asset():
-    from examples.surfer_memory_asset import main
+def test_register_upload_asset_surfer():
+    from examples.register_upload_asset_surfer import main
     main()