+++ tests/integration/test_use_cases/test_08_asset_registration.py
@@ -9,7 +9,7 @@
 import secrets
 import json
 
-from starfish.asset import MemoryAsset
+from starfish.asset import DataAsset
 from starfish.agent import SurferAgent
 from starfish.ddo.starfish_ddo import StarfishDDO
 
@@ -17,12 +17,12 @@ from starfish.ddo.starfish_ddo import StarfishDDO
 
 def test_08_asset_registration(resources, surfer_agent):
     test_data = secrets.token_hex(1024)
-    asset1 = MemoryAsset(data=test_data)
+    asset1 = DataAsset.create('TestAsset', test_data)
     listing1 = surfer_agent.register_asset(asset1, resources.listing_data)
     assert(listing1)
     assert(listing1.asset)
 
-    asset2 = MemoryAsset(data=test_data)
+    asset2 = DataAsset.create('TestAsset', test_data)
     listing2 = surfer_agent.register_asset(asset2, resources.listing_data)
     assert(listing2)
     assert(listing2.asset)