+++ tests/integration/test_use_cases/test_pd_case_asset_listing.py
@@ -11,10 +11,7 @@ import logging
 import json
 from web3 import Web3
 
-from starfish.asset import (
-    FileAsset,
-    RemoteAsset,
-)
+from starfish.asset import DataAsset
 
 def test_pd_case_file_transfer(ocean, config, resources, surfer_agent, squid_agent):
 
@@ -44,7 +41,7 @@ def test_pd_case_file_transfer(ocean, config, resources, surfer_agent, squid_age
         },
         'tags': [pd_test_case_tag],
     }
-    asset_sale = RemoteAsset(url=dummy_url)
+    asset_sale = DataAsset.create_from_url('TestAsset', dummy_url)
     # print('metadata ',squid_agent._convert_listing_asset_to_metadata(asset_sale, listing_data))
 
     listing = squid_agent.register_asset(asset_sale, listing_data, account=publisher_account)