+++ tests/unit/agent/test_squid_agent.py
@@ -8,12 +8,11 @@ import math
 
 from starfish.agent.squid_agent import SquidAgent
 from starfish.asset import (
-    Asset,
     BundleAsset,
-    RemoteAsset,
+    DataAsset,
 )
 from starfish.exceptions import StarfishPurchaseError
-from tests.unit.mocks.mock_squid_model import MockSquidModel
+from tests.unit.mocks.mock_squid_agent_adapter import MockSquidAgentAdapter
 
 
 VALID_DID = 'did:op:' + secrets.token_hex(64)
@@ -29,7 +28,7 @@ def _register_asset(ocean, resources, config):
     account = ocean.get_account(config.accounts[0].as_dict)
     agent = SquidAgent(ocean)
     assert(agent)
-    asset = RemoteAsset(url=resources.asset_remote)
+    asset = DataAsset.create_from_url('test url asset', resources.asset_remote)
     assert(asset)
     listing = agent.register_asset(asset, resources.listing_data, account)
     return (listing, agent, asset)
@@ -44,11 +43,11 @@ def test_init(ocean):
 
     agent = SquidAgent(ocean, TEST_INIT_PARMS)
     assert(agent)
-    assert isinstance(agent.squid_model, MockSquidModel)
-    assert(agent.squid_model.options['aquarius_url'] == TEST_INIT_PARMS['aquarius_url'])
-    assert(agent.squid_model.options['brizo_url'] == TEST_INIT_PARMS['brizo_url'])
-    assert(agent.squid_model.options['secret_store_url'] == TEST_INIT_PARMS['secret_store_url'])
-    assert(agent.squid_model.options['storage_path'] == TEST_INIT_PARMS['storage_path'])
+    assert isinstance(agent.get_adapter, MockSquidAgentAdapter)
+    assert(agent.get_adapter.options['aquarius_url'] == TEST_INIT_PARMS['aquarius_url'])
+    assert(agent.get_adapter.options['brizo_url'] == TEST_INIT_PARMS['brizo_url'])
+    assert(agent.get_adapter.options['secret_store_url'] == TEST_INIT_PARMS['secret_store_url'])
+    assert(agent.get_adapter.options['storage_path'] == TEST_INIT_PARMS['storage_path'])
 
     agent = SquidAgent(ocean,
         aquarius_url = TEST_INIT_PARMS['aquarius_url'],
@@ -57,10 +56,10 @@ def test_init(ocean):
         storage_path = TEST_INIT_PARMS['storage_path']
     )
     assert(agent)
-    assert(agent.squid_model.options['aquarius_url'] == TEST_INIT_PARMS['aquarius_url'])
-    assert(agent.squid_model.options['brizo_url'] == TEST_INIT_PARMS['brizo_url'])
-    assert(agent.squid_model.options['secret_store_url'] == TEST_INIT_PARMS['secret_store_url'])
-    assert(agent.squid_model.options['storage_path'] == TEST_INIT_PARMS['storage_path'])
+    assert(agent.get_adapter.options['aquarius_url'] == TEST_INIT_PARMS['aquarius_url'])
+    assert(agent.get_adapter.options['brizo_url'] == TEST_INIT_PARMS['brizo_url'])
+    assert(agent.get_adapter.options['secret_store_url'] == TEST_INIT_PARMS['secret_store_url'])
+    assert(agent.get_adapter.options['storage_path'] == TEST_INIT_PARMS['storage_path'])
 
 def test_register_asset(ocean, resources, config):
     listing, agent, asset = _register_asset(ocean, resources, config)
@@ -71,19 +70,26 @@ def test_register_bundle_asset(ocean, resources, config):
     account = ocean.get_account(config.accounts[0].as_dict)
     agent = SquidAgent(ocean)
     assert(agent)
-    bundle_asset = BundleAsset()
+    bundle_asset = BundleAsset.create('test bundle asset')
     for index in range(0, 5):
-        asset = RemoteAsset(url=resources.asset_remote)
+        asset = DataAsset.create_from_url('test remote asset', resources.asset_remote)
         assert(asset)
         bundle_asset.add(f'name_{index}', asset)
     listing = agent.register_asset(bundle_asset, resources.listing_data, account)
     assert(listing)
 
-def test_validate_asset(ocean, metadata):
+def test_register_asset_invalid_data(ocean, resources, config):
+    account = ocean.get_account(config.accounts[0].as_dict)
+    agent = SquidAgent(ocean)
+    assert(agent)
+    asset = DataAsset.create('test remote asset', 'some data that is never going to be saved')
+    with pytest.raises(ValueError):
+        listing = agent.register_asset(asset, resources.listing_data, account)
+
+def test_validate_asset(ocean):
     agent = SquidAgent(ocean, TEST_INIT_PARMS)
     assert(agent)
-#    metadata['base']['dateCreated'] = '2012-10-10T17:00:000Z'
-    asset = Asset(metadata)
+    asset = DataAsset.create_from_url('TestAsset', 'http://dex.sg')
     assert(agent.validate_asset(asset))
 
 def test_get_listing(ocean, resources, config):
@@ -108,7 +114,6 @@ def test_search_listings(ocean, resources, config):
 def test_purchase_asset(ocean, resources, config):
     listing, agent, asset = _register_asset(ocean, resources, config)
     account = ocean.get_account(config.accounts[1].as_dict)
-    print(account.ocean_balance)
     purchase = agent.purchase_asset(listing, account)
     assert(purchase)
 
@@ -136,7 +141,7 @@ def test_price_out_of_range(ocean, resources, config):
     account = ocean.get_account(config.accounts[0].as_dict)
     agent = SquidAgent(ocean)
     assert(agent)
-    asset = RemoteAsset(url=resources.asset_remote)
+    asset = DataAsset.create_from_url('test squid asset with url', resources.asset_remote)
     assert(asset)
     resources.listing_data['price'] = -1
     with pytest.raises(ValueError):