+++ tests/unit/asset/test_bundle_asset.py
@@ -7,19 +7,24 @@ import secrets
 import logging
 import json
 
-from starfish.asset import MemoryAsset
+from starfish.asset import DataAsset
 from starfish.asset import BundleAsset
 
 TEST_ASSET_COUNT = 4
 
 
-def test_init(metadata):
-    bundle = BundleAsset()
+def test_init():
+    bundle = BundleAsset.create('Bundle Asset')
     assert(bundle)
 
+    bundle = BundleAsset({
+        'name': 'Bundle Asset',
+        'type': 'bundle'
+    })
+    assert(bundle)
 
-def test_bundle_asset_add_access_remove(ocean, metadata, config):
-    bundle = BundleAsset()
+def test_bundle_asset_add_access_remove(ocean, config):
+    bundle = BundleAsset.create('name')
     asset_list = {}
 
     assert(bundle.is_bundle)
@@ -28,7 +33,7 @@ def test_bundle_asset_add_access_remove(ocean, metadata, config):
     for index in range(0, TEST_ASSET_COUNT):
         test_data = secrets.token_hex(1024)
         name = f'name_{index}'
-        asset_list[name] = MemoryAsset(metadata=metadata, data=test_data)
+        asset_list[name] = DataAsset.create(f'Asset_{index}', test_data)
         bundle.add(name, asset_list[name])
         assert(bundle[name])
         assert(bundle.asset_count == index + 1)