+++ tests/unit/asset/test_modules.py
@@ -9,30 +9,24 @@ import json
 
 from starfish.asset import (
     create_asset_from_metadata,
-    Asset,
     AssetBase,
     BundleAsset,
-    MemoryAsset,
+    DataAsset,
     OperationAsset,
-    RemoteAsset,
 )
 
 
-def test_create_asset_from_metadata(metadata):
-    asset = create_asset_from_metadata(metadata)
-    assert(asset)
-    assert(isinstance(asset, BundleAsset))
+def test_create_asset_from_metadata():
 
     type_list = [
         ( 'bundle', BundleAsset ),
-        ( 'data', MemoryAsset ),
-        ( 'memory', MemoryAsset ),
+        ( 'dataset', DataAsset ),
         ( 'operation', OperationAsset ),
-        ( 'unknown', Asset ),
     ]
     for type_name, class_type in type_list:
         metadata = {
-            'type': type_name
+            'type': type_name,
+            'name': f'TestAsset_{type_name}'
         }
         asset = create_asset_from_metadata(metadata)
         assert(asset)