+++ tests/unit/conftest.py
@@ -4,7 +4,7 @@ import secrets
 
 from tests.unit.libs.unit_test_config import unitTestConfig
 from starfish import Ocean
-from tests.unit.mocks.mock_squid_model import MockSquidModel
+from tests.unit.mocks.mock_squid_agent_adapter import MockSquidAgentAdapter
 
 
 
@@ -13,7 +13,7 @@ def ocean():
     result = Ocean(keeper_url=unitTestConfig.keeper_url,
             contracts_path=unitTestConfig.contracts_path,
             gas_limit=unitTestConfig.gas_limit,
-            squid_model_class=MockSquidModel,
+            squid_agent_adapter_class=MockSquidAgentAdapter,
             connect=False
     )
     return result