+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Change Log
 
+### Release v0.5.0
+
+*   Removed FileAsset, MemoryAsset, RemoteAsset. Replaced with DataAsset
+*   Cleanup documentation
+*   Renamed Squid/Surfer Models to Squid/Surfer Agent Adapters
+
 ### Release v0.4.11
 
 *   Add account balance and asset price check before purchasing an asset
@@ -23,6 +29,7 @@
 *   New squid-agent method watch_provider_events
 
 ### Release v0.4.6
+
 *   Fixed logging, so that only in test we setup the log level
 *   Add a test to see if we have already purchased an asset in squid
 *   Add MemoryAsset.save method